进一步拆分 SFTP 前端状态
